refactor[Qt event handler]: move Qt event handling into `event_handler.cpp`
